refactor: cleanups
