+++ packages/sdk-ts/src/client/indexer/ws/GrpcWebSocketCodec.ts
@@ -28,15 +28,15 @@ export const GrpcWebSocketCodec = {
 
   encodeTakerRequest(input: RFQRequestInputType): Uint8Array {
     const request = InjectiveRFQExchangeRpcPb.RFQRequestType.create({
+      margin: input.margin,
+      status: input.status,
       clientId: input.clientId,
       marketId: input.marketId,
-      direction: input.direction,
-      margin: input.margin,
       quantity: input.quantity,
+      direction: input.direction,
       worstPrice: input.worstPrice,
-      requestAddress: input.requestAddress,
       expiry: BigInt(input.expiry),
-      status: input.status,
+      requestAddress: input.requestAddress,
       // Server-side fields (set to 0/defaults for client requests)
       createdAt: 0n,
       updatedAt: 0n,
@@ -78,24 +78,24 @@ export const GrpcWebSocketCodec = {
 
   encodeMakerQuote(input: RFQQuoteType): Uint8Array {
     const quote = InjectiveRFQExchangeRpcPb.RFQQuoteType.create({
-      chainId: input.chainId,
-      contractAddress: input.contractAddress,
-      marketId: input.marketId,
-      rfqId: BigInt(input.rfqId),
-      takerDirection: input.takerDirection,
-      margin: input.margin,
-      quantity: input.quantity,
       price: input.price,
+      maker: input.maker,
+      taker: input.taker,
+      margin: input.margin,
+      status: input.status,
+      chainId: input.chainId,
+      marketId: input.marketId,
+      quantity: input.quantity,
+      rfqId: BigInt(input.rfqId),
+      signature: input.signature,
+      takerDirection: input.takerDirection,
+      contractAddress: input.contractAddress,
       expiry: {
         ...(input.expiry?.height && { height: BigInt(input.expiry.height) }),
         ...(input.expiry?.timestamp && {
           timestamp: BigInt(input.expiry.timestamp),
         }),
       },
-      maker: input.maker,
-      taker: input.taker,
-      signature: input.signature,
-      status: input.status,
       // Server-side fields (set to 0/defaults for client requests)
       createdAt: 0n,
       updatedAt: 0n,
