+++ packages/sdk-ts/src/client/indexer/ws/constants.ts
@@ -3,8 +3,8 @@ import type { WsReconnectConfig } from '../types'
 export const DEFAULT_RECONNECT_CONFIG: WsReconnectConfig = {
   enabled: true,
   maxAttempts: 10,
-  initialDelayMs: 1000,
   maxDelayMs: 30000,
+  initialDelayMs: 1000,
   backoffMultiplier: 2,
 }
