+++ packages/sdk-ts/src/client/indexer/ws/rfq/IndexerWsMakerStream.ts
@@ -1,16 +1,16 @@
+import { GrpcDecodeError } from '../../types'
 import { RFQ_GRPC_PATHS } from '../constants.js'
-import { GrpcDecodeError } from '../../types/ws.js'
 import { GrpcWebSocketCodec } from '../GrpcWebSocketCodec.js'
 import { GrpcWebSocketTransport } from '../GrpcWebSocketTransport.js'
 import { IndexerGrpcRfqTransformer } from '../../transformers/IndexerGrpcRfqTransformer.js'
+import type { WsState } from '../../types'
 import type { RFQQuoteType } from '../../types'
-import type { WsState } from '../../types/ws.js'
 import type {
   MakerStreamConfig,
   MakerStreamEvents,
   RFQStreamErrorData,
   RFQMakerStreamAckData,
-} from '../../types/rfq.js'
+} from '../../types'
 
 type MakerEventListener<T extends keyof MakerStreamEvents> = (
   data: MakerStreamEvents[T],
